🎨 좌표 관련 함수를 위한 CoordService 클래스 생성
